+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 CHANGELOG
 
 ## vNext
+- Added assertion for an exception to surface _immediately_, when retrieving an iterator.
+- Added assertion for an exception to surface _deferred_, when enumerating an iterator.